src/requirements.txt
tiktoken
uvicorn
fastapi
httpx
python-dotenv